refactor(OBS-RESOLVE): arch-compliance — split feedback repo, inject handler deps, dedup provider naming
