+++ apps/core/src/app/bootstrap/runtime-observer-feedback-message-action.ts
@@ -1,18 +1,11 @@
-import type { ObserverActivationStatus } from '../../config/settings/observer-activation.js';
 import type { ObserverDigestMessageView } from '../../domain/observer-digest-view.js';
 import { markObserverDigestInsight } from '../../domain/observer-digest-view.js';
 import type {
   ObserverInsightType,
   ObserverOwnerActionInsight,
   ObserverOwnerActionResult,
+  ObserverVerifiedOwner,
 } from '../../domain/ports/observer-insights.js';
-import { getRuntimeSettingsForConfig } from '../../config/index.js';
-import { resolveVerifiedObserverActivationStatus } from '../../config/settings/observer-activation.js';
-import { getRuntimeStorage } from '../../adapters/storage/postgres/runtime-store.js';
-import { DEFAULT_MEMORY_APP_ID } from '../../memory/app-memory-boundaries.js';
-import { nowIso } from '../../shared/time/datetime.js';
-import { logger } from '../../infrastructure/logging/logger.js';
-import type { ChannelWiring } from './channel-wiring-types.js';
 import type { ObserverFeedbackAction } from '../../domain/message-actions.js';
 import type {
   MessageActionOutcome,
@@ -47,9 +40,10 @@ const FAILED_RECEIPT = 'This insight could not be updated.';
 export interface ObserverFeedbackMessageActionDeps {
   appId: string;
   nowIso: () => string;
-  // The CURRENT DB-verified owner route (resolveVerifiedObserverActivationStatus);
-  // an unverified/unconfigured owner yields a status WITHOUT `owner`.
-  resolveVerifiedOwner: () => Promise<ObserverActivationStatus>;
+  // The CURRENT DB-verified owner route; an unverified/unconfigured owner
+  // yields a value WITHOUT `owner`. Injected by the wiring so this handler
+  // never imports config/adapters.
+  resolveVerifiedOwner: () => Promise<ObserverVerifiedOwner>;
   findInsightForOwnerAction: (input: {
     appId: string;
     recipient: string;
@@ -113,11 +107,10 @@ export async function handleObserverFeedbackAction(
   }
   try {
     // Step 2: the callback identity + route MUST be the current verified owner DM.
-    const activation = await deps.resolveVerifiedOwner();
-    if (!('owner' in activation)) {
+    const owner = (await deps.resolveVerifiedOwner()).owner;
+    if (!owner) {
       return { state: 'denied', receipt: OWNER_ONLY_RECEIPT };
     }
-    const owner = activation.owner;
     if (
       action.userId !== owner.recipient ||
       action.conversationJid !== owner.conversationJid ||
@@ -203,43 +196,4 @@ export async function handleObserverFeedbackAction(
     // unhandled rejection escaping into the channel adapter.
     return { state: 'invalid', receipt: FAILED_RECEIPT };
   }
-}
-
-export function registerRuntimeObserverFeedbackMessageAction(
-  channelWiring: ChannelWiring,
-): void {
-  const deps: ObserverFeedbackMessageActionDeps = {
-    appId: DEFAULT_MEMORY_APP_ID,
-    nowIso: () => nowIso(),
-    resolveVerifiedOwner: () =>
-      resolveVerifiedObserverActivationStatus(
-        getRuntimeSettingsForConfig(),
-        DEFAULT_MEMORY_APP_ID,
-        getRuntimeStorage().repositories.conversations,
-      ),
-    findInsightForOwnerAction: (input) =>
-      getRuntimeStorage().repositories.observerInsights.findInsightForOwnerAction(
-        input,
-      ),
-    applyOwnerAction: (input) =>
-      getRuntimeStorage().repositories.observerInsights.applyOwnerAction(input),
-    // Re-load the durable digest view by the callback's own (recipient, localDay)
-    // — the button carries the day it was rendered for, so this pins to the exact
-    // delivery and is timezone-stable (no surfacedAt/timezone recompute).
-    loadReservationView: async ({ recipient, localDay }) => {
-      const reservation =
-        await getRuntimeStorage().repositories.observerInsights.findDigestReservation(
-          { appId: DEFAULT_MEMORY_APP_ID, recipient, localDay },
-        );
-      return reservation?.renderedView ?? null;
-    },
-    listOwnerActions: (input) =>
-      getRuntimeStorage().repositories.observerInsights.listOwnerActionsForInsights(
-        input,
-      ),
-    warn: (context, message) => logger.warn(context, message),
-  };
-  channelWiring.setObserverFeedbackMessageActionHandler((action) =>
-    handleObserverFeedbackAction(deps, action),
-  );
 }
