+++ apps/core/src/app/bootstrap/runtime-services.ts
@@ -89,7 +89,7 @@ import {
 } from './runtime-services-active-compact.js';
 import { registerRuntimeLiveStopMessageAction } from './runtime-live-stop-message-action.js';
 import { registerRuntimeMemoryReviewMessageAction } from './runtime-memory-review-message-action.js';
-import { registerRuntimeObserverFeedbackMessageAction } from './runtime-observer-feedback-message-action.js';
+import { registerRuntimeObserverFeedbackMessageAction } from './runtime-observer-feedback-wiring.js';
 import { nowIso, nowMs, toIso } from '../../shared/time/datetime.js';
 import { LiveTurnAuthority } from '../../runtime/live-turn-authority.js';
 import type { LiveTurnRecoveryLoop } from '../../runtime/live-turn-recovery.js';
