+++ scripts/architecture-map.json
@@ -19,7 +19,7 @@
     "apps/core/src/adapters/storage/postgres/repositories/domain-repositories.postgres.ts": 2100,
     "apps/core/src/adapters/storage/postgres/repositories/control-plane-repository.postgres.ts": 760,
     "apps/core/src/jobs/ipc-skill-install-handlers.ts": 760,
-    "apps/core/src/app/bootstrap/channel-wiring.ts": 759,
+    "apps/core/src/app/bootstrap/channel-wiring.ts": 762,
     "apps/core/src/adapters/storage/postgres/repositories/canonical-job-repository.postgres.ts": 900,
     "apps/core/src/adapters/storage/postgres/repositories/live-turn-repository.postgres.ts": 770,
     "apps/core/src/adapters/storage/postgres/services/canonical-job-ops-service.ts": 807,
@@ -28,12 +28,12 @@
 
     "apps/core/src/app/bootstrap/inline-agent-loop-tools.ts": 745,
     "apps/core/src/app/bootstrap/runtime-app.ts": 725,
-    "apps/core/src/app/bootstrap/runtime-services.ts": 1080,
+    "apps/core/src/app/bootstrap/runtime-services.ts": 1120,
     "apps/core/src/channels/slack/channel-delivery-helpers.ts": 774,
     "apps/core/src/channels/slack/channel-delivery.ts": 732,
     "apps/core/src/channels/slack/channel-interactions.ts": 712,
     "apps/core/src/channels/telegram/channel-connect.ts": 708,
-    "apps/core/src/channels/telegram/channel-delivery.ts": 766,
+    "apps/core/src/channels/telegram/channel-delivery.ts": 770,
     "apps/core/src/cli/group.ts": 850,
     "apps/core/src/cli/model.ts": 723,
     "apps/core/src/config/settings/desired-state-service.ts": 830,
